requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0

# LangChain and AI components
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1

# Embeddings
sentence-transformers==2.2.2

# Vector store
chromadb==0.4.22
pysqlite3-binary==0.5.2

# Web scraping
beautifulsoup4==4.12.2
lxml==5.1.0

# Optional (if using Ollama)
# ollama==0.1.7

#pip install python-docx